requirements.txt
Flask
Jinja2
Werkzeug
requests
newsapi-python
gunicorn
Flask-Caching